+++ Common/src/main/java/fuzs/enchantinginfuser/api/world/item/enchantment/DefaultEnchantStatsProvider.java
@@ -1,6 +1,10 @@
 package fuzs.enchantinginfuser.api.world.item.enchantment;
 
+import fuzs.enchantinginfuser.core.ModCoreServices;
+import net.minecraft.core.BlockPos;
 import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.state.BlockState;
 
 /**
  * implementation for {@link EnchantStatsProvider} for vanilla values
@@ -20,6 +24,11 @@ public class DefaultEnchantStatsProvider implements EnchantStatsProvider {
     @Override
     public String[] getScalingNamespaces() {
         return new String[]{"minecraft"};
+    }
+
+    @Override
+    public float getEnchantPowerBonus(BlockState state, Level level, BlockPos pos) {
+        return ModCoreServices.ABSTRACTIONS.getEnchantPowerBonus(state, level, pos);
     }
 
     @Override