+++ Common/src/main/java/fuzs/enchantinginfuser/world/level/block/InfuserBlock.java
@@ -115,6 +115,8 @@ public class InfuserBlock extends EnchantmentTableBlock {
 
     @Override
     public void appendHoverText(ItemStack stack, @Nullable BlockGetter blockGetter, List<Component> list, TooltipFlag tooltipFlag) {
+        // don't let this go through when initially gathering tooltip data during start-up, configs do not exist then and it's ok if this is not searchable
+        if (!EnchantingInfuser.CONFIG.getHolder(ServerConfig.class).isAvailable()) return;
         Component component;
         if (this.type.config().allowModifyingEnchantments == ServerConfig.ModifyableItems.UNENCHANTED) {
             component = CHOOSE_TOOLTIP;