+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v4.0.3-1.19] - 2022-07-24
+### Added
+- Added support for bookshelves from Apotheosis, they will now provide enchanting power (Eterna) to the infuser
+- They will also allow the infuser to reach higher enchanting power values, depending on the type of bookshelf that is used
+- Added `ko_kr` translation by [Gyular], thanks!
+### Fixed
+- Fixed log spam due to server config not being loaded during start-up
+
 ## [v4.0.2-1.19] - 2022-07-24
 ### Changed
 - Made repair button more descriptive
@@ -19,3 +27,4 @@ The format is based on [Keep a Changelog].
 - Split into multi-loader project
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
+[Gyular]: https://github.com/gyular